Two files handTracking.py and handTrackingModule.py
Both perform similar tasks but handTrackingModule.py has a class that
can be imported later and used as a module in the future projects.
It performs following things
 - detectes hands from a video frame
 - draws the hands on to the screen
 - returns the position of the hand requested

While creating an object for the class, the parameters to set a few
default valuse don't work like they are supposed to work. IT is giving
out an error so the values are set to a few defaults and can't be
changed yet.